+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 2.2.7
 
 * [Phalcon] Added `services` part which can be used to `grabServiceFromContainer` and `addServiceToContainer` when conflicting module is used. By @sergeyklay
+* [Phalcon] **Refactored**. Moved in-memory session adapter to the separated namespace. By @sergeyklay
 
 #### 2.2.6 (October 2016)